pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
authors = ["Success Student <successstudent@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.2"
djangorestframework = "^3.13.1"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
djangorestframework-simplejwt = "^5.2.2"
python-magic = "^0.4.27"
Pillow = "^9.2.0"
drf-spectacular = "^0.24.2"
django-cors-headers = "^3.13.0"
django-allauth = "^0.51.0"
dj-rest-auth = "^2.2.5"
django-environ = "^0.9.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
